reorg native controller calls
