+++ spot-client/src/common/app-state/index.js
@@ -62,7 +62,6 @@ export * from './remote-control-service/selectors';
 export * from './modal/selectors';
 export * from './route/selectors';
 export * from './setup/selectors';
-export * from './spot-tv/middleware';
 export * from './spot-tv/selectors';
 export * from './wired-screenshare/selectors';
