--- spot-client/src/spot-tv/native-functions/bt-beacon/index.js
@@ -1,1 +1,0 @@
-import './middleware';
