+++ spot-client/src/spot-tv/native-functions/index.js
@@ -1,3 +1,2 @@
-export * from './bt-beacon';
 export * from './exit';
 export * from './volume-control';
