+++ spot-client/src/spot-tv/native-functions/native-controller/middleware.js
@@ -1,4 +1,5 @@
 
+import { SPOT_TV_SET_REMOTE_JOIN_CODE, SPOT_TV_SET_STATE } from 'common/app-state';
 import { BOOTSTRAP_COMPLETE } from 'common/app-state/bootstrap';
 import { MiddlewareRegistry } from 'common/redux';
 
@@ -17,7 +18,28 @@ MiddlewareRegistry.register(() => next => action => {
     case BOOTSTRAP_COMPLETE:
         nativeController._sendSpotClientReady();
         break;
+    case SPOT_TV_SET_REMOTE_JOIN_CODE:
+        nativeController.sendMessage('updateJoinCode', {
+            remoteJoinCode: action.remoteJoinCode
+        });
+        break;
+    case SPOT_TV_SET_STATE:
+        updateAPIMutedState(action.newState);
+        break;
     }
 
     return result;
 });
+
+/**
+ * Sends part of the new state through the external API if necessary.
+ *
+ * @param {Object} newState - The new state.
+ * @returns {void}
+ */
+function updateAPIMutedState(newState) {
+    const { audioMuted, videoMuted } = newState;
+
+    audioMuted !== undefined && nativeController.sendMessage('audioMutedState', audioMuted);
+    videoMuted !== undefined && nativeController.sendMessage('videoMutedState', videoMuted);
+}
