+++ spot-electron/main.js
@@ -13,7 +13,6 @@ const { createApplicationWindow } = require('./src/application-window');
 require('./src/spot-client-log-transport');
 require('./src/application-menu');
 require('./src/auto-updater');
-require('./src/bt-beacon');
 require('./src/client-control');
 require('./src/exit');
 require('./src/volume-control');
